\name{TheRSoftware-package}
\alias{TheRSoftware-package}
\alias{TheRSoftware}
\docType{package}
\title{Package illustrating the book: The R Software} 
\description{
This package enables one to use some functions used in the
book:\emph{The R Software, Fundamentals of Programming and Statistical
  Analysis}. One can also find the datasets used in the book. 
}
\details{
\tabular{ll}{
Package: \tab TheRSoftware\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-04\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Lafaye de Micheaux Pierre <lafaye@dms.umontreal.ca>, Remy Drouilhet <Remy.Drouilhet@upmf-grenoble.fr>, Liquet Benoit <benoit.liquet@isped.u-bordeaux2.fr>
}
\references{Book: The R Software, Fundamentals of Programming and Statistical
  Analysis}
\keyword{package}


